Fix transformer reconfigure (#18)
Only the initial sink/transformer subscribe RPC resulted in sink_config to be called. Therefore, the internal metadata object might be outdated, then new input_metrics were requested in a reconfigure. To fix that, this PR makes the sink_config method reentry save (in safe cases, otherwise exceptions will be thrown). I've tested it with the combinator and metricq-summary (a drain).